Move pulp_smash/tests/pulp3/{constants,utils}.py
Move the contents of these modules to
`pulp_smash/pulp3/{constants,utils}.py`. This makes it easier to blow
away the `pulp_smash/tests/` directory. Leave the old modules in place
with wildcard imports referencing the new module locations for the sake
of compatibility.

See: https://github.com/PulpQE/pulp-smash/issues/996
