+++ docs/api/pulp_smash.pulp3.constants.rst
@@ -1,0 +1,6 @@
+`pulp_smash.pulp3.constants`
+============================
+
+Location: :doc:`/index` → :doc:`/api` → :doc:`/api/pulp_smash.pulp3.constants`
+
+.. automodule:: pulp_smash.pulp3.constants
