+++ docs/api/pulp_smash.pulp3.rst
@@ -1,0 +1,6 @@
+`pulp_smash.pulp3`
+==================
+
+Location: :doc:`/index` → :doc:`/api` → :doc:`/api/pulp_smash.pulp3`
+
+.. automodule:: pulp_smash.pulp3
