+++ docs/api/pulp_smash.pulp3.utils.rst
@@ -1,0 +1,6 @@
+`pulp_smash.pulp3.utils`
+========================
+
+Location: :doc:`/index` → :doc:`/api` → :doc:`/api/pulp_smash.pulp3.utils`
+
+.. automodule:: pulp_smash.pulp3.utils
