+++ pulp_smash/pulp3/__init__.py
@@ -1,0 +1,2 @@
+# coding=utf-8
+"""Tools for Pulp 2 tests."""
