+++ pulp_smash/tests/pulp3/file/api_v3/test_crud_content_unit.py
@@ -8,14 +8,14 @@ from requests.exceptions import HTTPError
 
 from pulp_smash import api, config, selectors, utils
 from pulp_smash.constants import FILE_FEED_URL, FILE_URL
-from pulp_smash.tests.pulp3.constants import (
+from pulp_smash.pulp3.constants import (
     ARTIFACTS_PATH,
     FILE_CONTENT_PATH,
     FILE_REMOTE_PATH,
     REPO_PATH,
 )
 from pulp_smash.tests.pulp3.file.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.utils import (
+from pulp_smash.pulp3.utils import (
     delete_orphans,
     gen_remote,
     gen_repo,
