+++ pulp_smash/tests/pulp3/file/api_v3/test_orphans.py
@@ -10,7 +10,7 @@ from pulp_smash.constants import (
     FILE_FEED_URL,
 )
 from pulp_smash.exceptions import CalledProcessError
-from pulp_smash.tests.pulp3.constants import (
+from pulp_smash.pulp3.constants import (
     ARTIFACTS_PATH,
     FILE_CONTENT_PATH,
     FILE_REMOTE_PATH,
@@ -18,7 +18,7 @@ from pulp_smash.tests.pulp3.constants import (
 )
 
 from pulp_smash.tests.pulp3.file.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.utils import (
+from pulp_smash.pulp3.utils import (
     delete_orphans,
     delete_version,
     gen_remote,
