+++ pulp_smash/tests/pulp3/file/api_v3/test_repo_version.py
@@ -13,7 +13,7 @@ from pulp_smash.constants import (
     FILE_FEED_URL,
     FILE_LARGE_FEED_URL,
 )
-from pulp_smash.tests.pulp3.constants import (
+from pulp_smash.pulp3.constants import (
     FILE_CONTENT_PATH,
     FILE_REMOTE_PATH,
     FILE_PUBLISHER_PATH,
@@ -22,7 +22,7 @@ from pulp_smash.tests.pulp3.constants import (
 from pulp_smash.tests.pulp3.file.api_v3.utils import gen_publisher
 from pulp_smash.tests.pulp3.file.utils import populate_pulp
 from pulp_smash.tests.pulp3.file.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.utils import (
+from pulp_smash.pulp3.utils import (
     delete_version,
     gen_remote,
     gen_repo,
