+++ pulp_smash/tests/pulp3/file/api_v3/test_sync.py
@@ -10,9 +10,9 @@ from pulp_smash.constants import (
     FILE_FEED_URL,
     FILE_LARGE_FEED_URL,
 )
-from pulp_smash.tests.pulp3.constants import FILE_REMOTE_PATH, REPO_PATH
+from pulp_smash.pulp3.constants import FILE_REMOTE_PATH, REPO_PATH
 from pulp_smash.tests.pulp3.file.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.utils import (
+from pulp_smash.pulp3.utils import (
     gen_remote,
     gen_repo,
     get_auth,
