+++ pulp_smash/tests/pulp3/file/api_v3/test_unlinking_repo.py
@@ -6,14 +6,14 @@ from urllib.parse import urljoin
 
 from pulp_smash import api, config, selectors, utils
 from pulp_smash.constants import FILE_FEED_URL
-from pulp_smash.tests.pulp3.constants import (
+from pulp_smash.pulp3.constants import (
     FILE_REMOTE_PATH,
     FILE_PUBLISHER_PATH,
     REPO_PATH,
 )
 from pulp_smash.tests.pulp3.file.api_v3.utils import gen_publisher
 from pulp_smash.tests.pulp3.file.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.utils import (
+from pulp_smash.pulp3.utils import (
     gen_remote,
     gen_repo,
     get_auth,
