+++ pulp_smash/tests/pulp3/pulpcore/api_v3/test_auth.py
@@ -11,9 +11,9 @@ from requests.auth import HTTPBasicAuth
 from requests.exceptions import HTTPError
 
 from pulp_smash import api, config, selectors, utils
-from pulp_smash.tests.pulp3.constants import BASE_PATH, JWT_PATH, USER_PATH
+from pulp_smash.pulp3.constants import BASE_PATH, JWT_PATH, USER_PATH
 from pulp_smash.tests.pulp3.pulpcore.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.utils import JWTAuth
+from pulp_smash.pulp3.utils import JWTAuth
 
 
 class AuthTestCase(unittest.TestCase, utils.SmokeTest):
