+++ pulp_smash/tests/pulp3/pulpcore/api_v3/test_crd_artifacts.py
@@ -9,9 +9,9 @@ from requests.exceptions import HTTPError
 from pulp_smash import api, cli, config, utils
 from pulp_smash.constants import FILE_URL
 from pulp_smash.exceptions import CalledProcessError
-from pulp_smash.tests.pulp3.constants import ARTIFACTS_PATH
+from pulp_smash.pulp3.constants import ARTIFACTS_PATH
 from pulp_smash.tests.pulp3.pulpcore.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.utils import delete_orphans, get_auth
+from pulp_smash.pulp3.utils import delete_orphans, get_auth
 
 
 class ArtifactTestCase(unittest.TestCase, utils.SmokeTest):
