+++ pulp_smash/tests/pulp3/pulpcore/api_v3/test_crud_repos.py
@@ -5,9 +5,9 @@ import unittest
 from requests.exceptions import HTTPError
 
 from pulp_smash import api, config, selectors, utils
-from pulp_smash.tests.pulp3.constants import REPO_PATH
+from pulp_smash.pulp3.constants import REPO_PATH
 from pulp_smash.tests.pulp3.pulpcore.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.utils import gen_repo, get_auth
+from pulp_smash.pulp3.utils import gen_repo, get_auth
 
 
 class CRUDRepoTestCase(unittest.TestCase, utils.SmokeTest):
