+++ pulp_smash/tests/pulp3/pulpcore/api_v3/test_status.py
@@ -6,7 +6,7 @@ from jsonschema import validate
 from requests.exceptions import HTTPError
 
 from pulp_smash import api, config, utils
-from pulp_smash.tests.pulp3.constants import STATUS_PATH
+from pulp_smash.pulp3.constants import STATUS_PATH
 from pulp_smash.tests.pulp3.pulpcore.utils import set_up_module as setUpModule  # pylint:disable=unused-import
 
 STATUS = {
