+++ pulp_smash/tests/pulp3/pulpcore/api_v3/test_workers.py
@@ -7,9 +7,9 @@ from random import choice
 from requests.exceptions import HTTPError
 
 from pulp_smash import api, config, selectors, utils
-from pulp_smash.tests.pulp3.constants import WORKER_PATH
+from pulp_smash.pulp3.constants import WORKER_PATH
 from pulp_smash.tests.pulp3.pulpcore.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.utils import get_auth
+from pulp_smash.pulp3.utils import get_auth
 
 _DYNAMIC_WORKER_ATTRS = ('last_heartbeat',)
 """Worker attributes that are dynamically set by Pulp, not set by a user."""
